requirements.txt
pandas
python-docx
pyqt5
tkinter